// src/typings/system.d.ts
declare namespace I18nType {
  type langType = 'en' | 'zh-CN' | 'km-KH';

  interface Schema {
    system: {
      title: string;
    };
    routes: {
      dashboard: {
        analysis: string;
      };
      heat: {
        homepage: string;
      };
      track: {
        navigation: string;
      };
      nitrogen: {
        navigation: string;
      };
      count: {
        navigation: string;
      };
      product: {
        navigation: string;
      };
      vacuum: {
        homepage: string;
      };
      system: {
        _value: string;
        navigation: string;
        oil: string;
        maintain: string;
      };
      log: {
        _value: string;
        output: string;
      };
      alarm: {
        _value: string;
        navigation: string;
      };
      profile: {
        _value: string;
        heat: string;
        track: string;
        nitrogen: string;
        vacuum: string;
      };
      user: {
        _value: string;
        home: string;
        accredit: string;
        other: string;
      };
      ipconfig: {
        _value: string;
        homepage: string;
      };
      parameter: {
        _value: string;
        furnace: string;
        system: string;
      };
      file: {
        _value: string;
        heat: string;
        track: string;
        n2: string;
        vacuum: string;
        mes: string;
      };
      about: string;
    };
  }
}